pyproject.toml
[project]
name = "ciso-aac"
version = "0.1.0"
description = "A client library for accessing CISO Assistant API"
authors = [
    { name = "Johan Lundberg", email = "lundberg@sunet.se" }
]
readme = "README.md"
license = "BSD-2-clause"
license-files = ["LICEN[CS]E*"]

classifiers = [
    "Environment :: Web Environment",
    "Framework :: AsyncIO",
    "Framework :: Trio",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: BSD License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Internet :: WWW/HTTP",
]

requires-python = ">=3.10"
dependencies = [
    "attrs>=25.3.0",
    "httpx>=0.28.1",
    "python-dateutil>=2.9.0.post0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.sdist]
include = ["/src/ciso_aac"]
sources = ["src"]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["F", "I", "UP"]

[tool.ruff.lint.isort]
combine-as-imports = true

[dependency-groups]
dev = [
    "mypy>=1.15.0",
    "ruff>=0.11.2",
]

[tool.mypy]
ignore_missing_imports = true
strict = true

[project.urls]
Homepage = "https://github.com/SUNET/python-ciso-aac"
Issues = "https://github.com/SUNET/python-ciso-aac/issues"
